# src/api/docs/swagger.yaml
basePath: /
definitions:
  handler.APIResponse:
    properties:
      message:
        description: need fill only if error occurred
        type: string
    type: object
...
  handler.ListReceiptDirectionAPIResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/receipt.ReceiptDirection'
        type: array
      message:
        description: need fill only if error occurred
        type: string
    type: object
...
  handler.ReceiptDirectionAPIResponse:
    properties:
      item:
        $ref: '#/definitions/receipt.ReceiptDirection'
        type: object
      message:
        description: need fill only if error occurred
        type: string
    type: object
...
  receipt.ReceiptDirection:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      receipt_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  services.CreateReceiptDirectionRequest:
    properties:
      description:
        description: (required)
        maxLength: 255
        minLength: 3
        type: string
    required:
    - description
    type: object
...
  services.UpdateReceiptDirectionRequest:
    properties:
      description:
        description: (required)
        maxLength: 255
        minLength: 3
        type: string
    required:
    - description
    type: object
...
host: api.food.test
info:
  contact: {}
  description: This is Swagger docs for Food API
  license: {}
  title: Food API
  version: "1.0"
paths:
...
  /v1/receipts/{id}/directions:
    get:
      description: find receipt directions by params
      parameters:
      - description: Receipt id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ListReceiptDirectionAPIResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get receipt directions
      tags:
      - receipts
    post:
      parameters:
      - description: Receipt id
        in: path
        name: id
        required: true
        type: integer
      - description: params
        in: body
        name: ingredient
        required: true
        schema:
          $ref: '#/definitions/services.CreateReceiptDirectionRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ReceiptDirectionAPIResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create receipt direction
      tags:
      - receipts
  /v1/receipts/{id}/directions/{direction_id}:
    delete:
      parameters:
      - description: Receipt id
        in: path
        name: id
        required: true
        type: integer
      - description: Receipt direction id
        in: path
        name: direction_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204": {}
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete receipt direction
      tags:
      - receipts
    put:
      parameters:
      - description: Receipt id
        in: path
        name: id
        required: true
        type: integer
      - description: Receipt direction id
        in: path
        name: direction_id
        required: true
        type: integer
      - description: params
        in: body
        name: direction
        required: true
        schema:
          $ref: '#/definitions/services.UpdateReceiptDirectionRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ReceiptDirectionAPIResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update receipt direction
      tags:
      - receipts
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"